Update 2014-02-23-bdferrsimpleblog.md
I was reminded in rendering this post that I cannot simply write out an HTML tag in the usual way if I want to refer to it rather than apply it to the post. I searched for and found the proper characters to get the "less than" and "greater than" signs to simply display rather than be interpreted as marking HTML elements.